% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pnorm.R
\name{plot_pnorm}
\alias{plot_pnorm}
\title{Plot the result of a pnorm function call}
\usage{
plot_pnorm(q = NULL, mean = 0, sd = 1, lower.tail = TRUE)
}
\arguments{
\item{q}{numeric; The quantile to evaluate the \code{pnorm}}

\item{mean}{numeric; The mean of the normal distribution}

\item{sd}{numeric; The standard deviation of the normal distribution}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]} otherwise, \eqn{P[X > x]}.}
}
\value{
ggplot object
}
\description{
Plot the result of the \code{pnorm} function. The syntax is the same as the \code{pnorm} function.
}
\examples{
plot_pnorm(0.5, 0, 1, lower.tail = TRUE)

}
