% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcorr.R
\name{rcorr}
\alias{rcorr}
\title{rcorr}
\usage{
rcorr(mux = 0, muy = 0, n, sdx = 1, sdy = 1, r)
}
\arguments{
\item{mux}{mean of the first variable (x), default to 0}

\item{muy}{mean of the second variable (y), default to 0}

\item{n}{sample size}

\item{sdx}{standard deviation of the first variable (x), default to 1}

\item{sdy}{standard deviation of the second variable (y), default to 1}

\item{r}{the correlation between x and y}
}
\value{
a dataframe
}
\description{
generate two correlated variables
}
\examples{
rcorr(n = 100, r = 0.6)
}
