% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsummary.R
\name{rsummary}
\alias{rsummary}
\title{rsummary}
\usage{
rsummary(x, na.rm = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{logical; remove NA values before applying the functions}
}
\value{
a named vector with summary statistics
}
\description{
a simple \code{summary()} function with extra statistics
}
\examples{
x <- rnorm(100)
rsummary(x)
}
